logout

+++ IoT/src/app/components/bottom-tab/bottom-tab.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
+import { HttpClient } from '@angular/common/http';
 
 @Component({
   selector: 'app-bottom-tab',
@@ -9,11 +10,12 @@ import { CommonModule } from '@angular/common';
   templateUrl: './bottom-tab.component.html',
   styleUrl: './bottom-tab.component.css'
 })
-export class BottomTabComponent implements OnInit{
+
+export class BottomTabComponent implements OnInit {
   authorities: string[] = [];
   menuItems: any[] = [];
 
-  constructor(public router: Router) {}
+  constructor(public router: Router, private http: HttpClient) { }
 
   ngOnInit(): void {
     const storedAuthorities = JSON.parse(sessionStorage.getItem('authorities') || '[]');
@@ -41,13 +43,45 @@ export class BottomTabComponent implements OnInit{
     }
   }
 
+  private getBaseUrl(): string {
+    const host = window.location.hostname;
+
+    if (host === 'localhost') {
+      return 'http://localhost:3000';
+    } else if (host === '192.168.1.82') {
+      return 'http://192.168.1.82:3000';
+    } else if (host === '192.168.1.14') {
+      return 'http://192.168.1.14:3000';
+    } else {
+      // Opción predeterminada
+      return 'http://localhost:3000';
+    }
+  }
+
   // Método para cerrar sesión
-  logout() {
-    // Elimina todos los elementos del sessionStorage
-    sessionStorage.clear();
-    
-    // Redirige a la página de inicio de sesión o a la página principal
-    this.router.navigate(['/login']);
+  logout(): void {
+    const username = sessionStorage.getItem('username'); // Asegúrate de que 'username' esté almacenado correctamente.
+
+    if (!username) {
+      console.error('No se encontró el username en sessionStorage');
+      this.router.navigate(['/login']);
+      return;
+    }
+
+    const baseUrl = this.getBaseUrl(); // Llama al método dentro de la clase
+    const logoutEndpoint = `${baseUrl}/logout`;
+
+    // Realiza la solicitud al endpoint correcto
+    this.http.post(logoutEndpoint, { username }).subscribe({
+      next: () => {
+        // Limpia el sessionStorage y redirige al usuario
+        sessionStorage.clear();
+        this.router.navigate(['/login']);
+      },
+      error: (error) => {
+        console.error('Error durante el logout:', error);
+      },
+    });
   }
 
   isActive(route: string): boolean {
