+++ Iot/src/app/components/side/side.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Router, ActivatedRoute } from '@angular/router';
 import { FormsModule } from '@angular/forms';
+import { HttpClient } from '@angular/common/http';
 
 @Component({
   selector: 'app-side',
@@ -24,86 +25,86 @@ export class SideComponent implements OnInit {
   // Menú original con las autoridades asociadas a cada elemento
   menuItemsTop = [
     { icon: 'home', label: 'Home', route: '/', submenu: [], isSubmenuOpen: false, authorities: [] },
-    { 
-      icon: 'add', 
-      label: 'Registrar', 
-      isSubmenuOpen: false, 
+    {
+      icon: 'add',
+      label: 'Registrar',
+      isSubmenuOpen: false,
       submenu: [
-        { label: 'Registrar Administrador', route: '/admin/create', authorities: ['create_users','super_administrador', 'administrador'] },
-        { label: 'Control Administrador', route: '/admin/index', authorities: ['list_users','super_administrador', 'administrador'] },
-        { label: 'Administradores Eliminados', route: '/admin/indexE', authorities: ['delete_user','super_administrador', 'administrador'] },
+        { label: 'Registrar Administrador', route: '/admin/create', authorities: ['create_users', 'super_administrador', 'administrador'] },
+        { label: 'Control Administrador', route: '/admin/index', authorities: ['list_users', 'super_administrador', 'administrador'] },
+        { label: 'Administradores Eliminados', route: '/admin/indexE', authorities: ['delete_user', 'super_administrador', 'administrador'] },
       ],
-      authorities: ['create_users', 'list_users', 'delete_user','super_administrador', 'administrador']
+      authorities: ['create_users', 'list_users', 'delete_user', 'super_administrador', 'administrador']
     },
-    { 
-      icon: 'notifications', 
-      label: 'Alertas', 
-      isSubmenuOpen: false, 
+    {
+      icon: 'notifications',
+      label: 'Alertas',
+      isSubmenuOpen: false,
       submenu: [
-        { label: 'Registrar Alertas', route: '/alert/create', authorities: ['create_alert','super_administrador', 'administrador'] },
-        { label: 'Control de Alertas', route: '/alert/index', authorities: ['list_alert','super_administrador', 'administrador'] },
-        { label: 'Alertas Eliminadas', route: '/alert/indexE', authorities: ['delete_alert','super_administrador', 'administrador'] },
+        { label: 'Registrar Alertas', route: '/alert/create', authorities: ['create_alert', 'super_administrador', 'administrador'] },
+        { label: 'Control de Alertas', route: '/alert/index', authorities: ['list_alert', 'super_administrador', 'administrador'] },
+        { label: 'Alertas Eliminadas', route: '/alert/indexE', authorities: ['delete_alert', 'super_administrador', 'administrador'] },
       ],
-      authorities: ['create_alert', 'list_alert', 'delete_alert','super_administrador', 'administrador']
+      authorities: ['create_alert', 'list_alert', 'delete_alert', 'super_administrador', 'administrador']
     },
-    { 
-      icon: 'subscriptions', 
-      label: 'Suscripciones', 
-      isSubmenuOpen: false, 
+    {
+      icon: 'subscriptions',
+      label: 'Suscripciones',
+      isSubmenuOpen: false,
       submenu: [
-        { label: 'Registrar Suscripciones', route: '/suscription/create', authorities: ['create_suscription','super_administrador', 'administrador'] },
-        { label: 'Control de Suscripciones', route: '/suscription/index', authorities: ['list_suscriptions','super_administrador', 'administrador'] },
-        { label: 'Suscripciones Eliminadas', route: '/suscription/indexE', authorities: ['delete_suscription','super_administrador', 'administrador'] },
+        { label: 'Registrar Suscripciones', route: '/suscription/create', authorities: ['create_suscription', 'super_administrador', 'administrador'] },
+        { label: 'Control de Suscripciones', route: '/suscription/index', authorities: ['list_suscriptions', 'super_administrador', 'administrador'] },
+        { label: 'Suscripciones Eliminadas', route: '/suscription/indexE', authorities: ['delete_suscription', 'super_administrador', 'administrador'] },
       ],
-      authorities: ['create_suscription', 'list_suscriptions', 'delete_suscription','super_administrador', 'administrador']
+      authorities: ['create_suscription', 'list_suscriptions', 'delete_suscription', 'super_administrador', 'administrador']
     },
-    { 
-      icon: 'apps', 
-      label: 'Servicios', 
-      isSubmenuOpen: false, 
+    {
+      icon: 'apps',
+      label: 'Servicios',
+      isSubmenuOpen: false,
       submenu: [
-        { label: 'Registrar Servicio', route: '/services/create', authorities: ['create_iot_service','super_administrador', 'administrador'] },
-        { label: 'Control de Servicios', route: '/services/index', authorities: ['list_iot_service','super_administrador', 'administrador'] },
-        { label: 'Servicios Eliminados', route: '/services/indexE', authorities: ['delete_iot_service','super_administrador', 'administrador'] },
+        { label: 'Registrar Servicio', route: '/services/create', authorities: ['create_iot_service', 'super_administrador', 'administrador'] },
+        { label: 'Control de Servicios', route: '/services/index', authorities: ['list_iot_service', 'super_administrador', 'administrador'] },
+        { label: 'Servicios Eliminados', route: '/services/indexE', authorities: ['delete_iot_service', 'super_administrador', 'administrador'] },
       ],
-      authorities: ['create_iot_service', 'list_iot_service', 'delete_iot_service','super_administrador', 'administrador']
+      authorities: ['create_iot_service', 'list_iot_service', 'delete_iot_service', 'super_administrador', 'administrador']
     },
-    { 
-      icon: 'sensors', 
-      label: 'Sensores', 
-      isSubmenuOpen: false, 
+    {
+      icon: 'sensors',
+      label: 'Sensores',
+      isSubmenuOpen: false,
       submenu: [
-        { label: 'Registrar Sensor', route: '/sensors/create', authorities: ['create_sensors','super_administrador', 'administrador'] },
-        { label: 'Control Sensores', route: '/sensors/index', authorities: ['list_sensors','super_administrador', 'administrador'] },
-        { label: 'Sensores Eliminados', route: '/sensors/indexE', authorities: ['delete_sensors','super_administrador', 'administrador'] },
+        { label: 'Registrar Sensor', route: '/sensors/create', authorities: ['create_sensors', 'super_administrador', 'administrador'] },
+        { label: 'Control Sensores', route: '/sensors/index', authorities: ['list_sensors', 'super_administrador', 'administrador'] },
+        { label: 'Sensores Eliminados', route: '/sensors/indexE', authorities: ['delete_sensors', 'super_administrador', 'administrador'] },
       ],
-      authorities: ['create_sensors', 'list_sensors', 'delete_sensors','super_administrador', 'administrador']
+      authorities: ['create_sensors', 'list_sensors', 'delete_sensors', 'super_administrador', 'administrador']
     },
-    { 
-      icon: 'history', 
-      label: 'Ver Historial', 
-      isSubmenuOpen: false, 
+    {
+      icon: 'history',
+      label: 'Ver Historial',
+      isSubmenuOpen: false,
       submenu: [
         { label: 'Control Sesiones', route: '/sessions/index', authorities: ['super_administrador'] },
         { label: 'Sesiones Eliminadas', route: '/sessions/indexE', authorities: ['super_administrador'] },
       ],
       authorities: ['super_administrador']
     },
-    { 
-      icon: 'people', 
-      label: 'Control Usuarios', 
-      isSubmenuOpen: false, 
+    {
+      icon: 'people',
+      label: 'Control Usuarios',
+      isSubmenuOpen: false,
       submenu: [
         { label: 'Control Usuarios', route: '/users/index', authorities: ['super_administrador'] },
       ],
       authorities: ['super_administrador']
     },
-    { 
-      icon: 'lock', 
-      label: 'Cambiar Contraseña', 
-      route: '/users/cambiarContra', 
-      submenu: [], 
-      isSubmenuOpen: false, 
+    {
+      icon: 'lock',
+      label: 'Cambiar Contraseña',
+      route: '/users/cambiarContra',
+      submenu: [],
+      isSubmenuOpen: false,
       authorities: []
     },
   ];
@@ -114,7 +115,7 @@ export class SideComponent implements OnInit {
 
   authorities: string[] = [];
 
-  constructor(private router: Router, private activatedRoute: ActivatedRoute) {}
+  constructor(private router: Router, private activatedRoute: ActivatedRoute, private http: HttpClient) { }
 
   ngOnInit(): void {
     // Detectar ruta activa
@@ -204,13 +205,47 @@ export class SideComponent implements OnInit {
     return this.currentRoute === route;
   }
 
-  logout(): void {
-    // Eliminar todos los datos de sessionStorage
-    sessionStorage.clear();
-    console.log("Sesión cerrada y sessionStorage limpiado.");
-    // Redirigir a la página de inicio de sesión
-    this.router.navigate(['/login']);
+  private getBaseUrl(): string {
+    const host = window.location.hostname;
+
+    if (host === 'localhost') {
+      return 'http://localhost:3000';
+    } else if (host === '192.168.1.82') {
+      return 'http://192.168.1.82:3000';
+    } else if (host === '192.168.1.14') {
+      return 'http://192.168.1.14:3000';
+    } else {
+      // Opción predeterminada
+      return 'http://localhost:3000';
+    }
   }
+
+  // Método para cerrar sesión
+  logout(): void {
+    const username = sessionStorage.getItem('username'); // Asegúrate de que 'username' esté almacenado correctamente.
+
+    if (!username) {
+      console.error('No se encontró el username en sessionStorage');
+      this.router.navigate(['/login']);
+      return;
+    }
+
+    const baseUrl = this.getBaseUrl(); // Llama al método dentro de la clase
+    const logoutEndpoint = `${baseUrl}/logout`;
+
+    // Realiza la solicitud al endpoint correcto
+    this.http.post(logoutEndpoint, { username }).subscribe({
+      next: () => {
+        // Limpia el sessionStorage y redirige al usuario
+        sessionStorage.clear();
+        this.router.navigate(['/login']);
+      },
+      error: (error) => {
+        console.error('Error durante el logout:', error);
+      },
+    });
+  }
+
   navigateTo(route: string | undefined): void {
     if (route) {
       this.router.navigate([route]);
@@ -218,5 +253,5 @@ export class SideComponent implements OnInit {
       console.error('Route is undefined');
     }
   }
-  
+
 }
